:wrench: Simplify DNS configuration for 'demo' in test environment
In this commit, we simplified the DNS configuration file for the 'test-dns.service.justice.gov.uk' service. Specifically, we removed the A record for 'test-dns.service.justice.gov.uk.', which was using a placeholder IP. We retained and rearranged the details of the 'demo' CNAME record to enhance readability. The TTL remains unchanged at '300'. The change aims to streamline the DNS configuration process and remove unused records.