:wrench: Update DNS record type from A to CNAME
This commit alters the DNS record type for the test-dns.service.justice.gov.uk configuration. We're moving from an 'A' (address) record, which points a hostname to an IP address, to a 'CNAME' (Canonical Name) record, which points a hostname to another domain name. The target for the new CNAME record is example.com. This change will enable us to manage and route traffic more dynamically.